Code smells, reorganizing, using TileTypeEnum, adding tests
