+++ clean-sweep/src/main/java/com/group9/cleansweep/InitFloorPlan.java
@@ -1,4 +1,4 @@
-package com.group9.cleansweep.controlsystem.floorplanfile;
+package com.group9.cleansweep;
 
 import java.io.FileInputStream;
 import java.io.IOException;
