+++ clean-sweep/src/main/java/com/group9/cleansweep/enums/TileTypeEnum.java
