--- control-system/src/main/java/com/group9/cleansweep/enums/FloorPlanTypeEnum.java
@@ -1,5 +1,0 @@
-package com.group9.cleansweep.enums;
-
-public enum FloorPlanTypeEnum {
-	BARE_FOOT, LOW_PILE_CARPET, HIGH_PILE_CARPET
-}
